Исправить мелкие неточности.
