+++ src/api/handlers/BaseHttpHandler.java
@@ -1,10 +1,11 @@
-package api.habdlers;
+package api.handlers;
 
 import api.adapters.DurationAdapter;
 import api.adapters.LocalDateTimeAdapter;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpExchange;
+import com.sun.net.httpserver.HttpHandler;
 import manager.task.TaskManager;
 
 import java.io.IOException;
@@ -13,11 +14,11 @@ import java.nio.charset.StandardCharsets;
 import java.time.Duration;
 import java.time.LocalDateTime;
 
-public class BaseHttpHandler {
+public abstract class BaseHttpHandler implements HttpHandler {
 
     protected final TaskManager taskManager;
 
-    protected Gson gson = new GsonBuilder()
+    protected final Gson gson = new GsonBuilder()
             .registerTypeAdapter(Duration.class, new DurationAdapter())
             .registerTypeAdapter(LocalDateTime.class, new LocalDateTimeAdapter())
             .setPrettyPrinting()
