+++ src/api/handlers/Endpoints.java
@@ -1,0 +1,10 @@
+package api.handlers;
+
+public enum Endpoints {
+    GET_ALL,
+    GET_ID,
+    GET_ALL_SUBTASKS,
+    POST_CREATE,
+    DELETE_ID,
+    UNKNOWN
+}
