+++ src/api/handlers/HistoryHandler.java
@@ -1,12 +1,11 @@
-package api.habdlers;
+package api.handlers;
 
 import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpHandler;
 import manager.task.TaskManager;
 
 import java.io.IOException;
 
-public class HistoryHandler extends BaseHttpHandler implements HttpHandler {
+public class HistoryHandler extends BaseHttpHandler {
 
     public HistoryHandler(TaskManager taskManager) {
         super(taskManager);
@@ -14,6 +13,8 @@ public class HistoryHandler extends BaseHttpHandler implements HttpHandler {
 
     @Override
     public void handle(HttpExchange httpExchange) throws IOException {
-        sendText(httpExchange, gson.toJson(taskManager.getHistory()));
+        if (httpExchange.getRequestMethod().equals("GET")) {
+            sendText(httpExchange, gson.toJson(taskManager.getHistory()));
+        }
     }
 }
