+++ test/api/handlers/TasksHandlerTest.java
@@ -1,4 +1,4 @@
-package api.habdlers;
+package api.handlers;
 
 import api.ApiTestMethods;
 import api.HttpTaskServer;
